DOC/DEP: opencv py3.6 inst. note :D, rm viz extra
the reason for removing the 'viz' extra is that, as the last
hour's fiasco showed, just not installing opencv-python will mean
that we can't use wotplot as is. i'd need to do some extra engineering
to support this -- probably some try/except statements or something.
See https://stackoverflow.com/a/27361558.

this is doable but a lot of work, and also now it's pretty easy to
install matplotlib and opencv-python. So! nbd for now