DOC/MNT: small README / tutorial improvements
also removed the code that sets all tick_params() to True for some
reason -- fr what was i doing with that?? i think removing this
slightly changed some of the visualizations of small dot plot
matrices (makes sense -- i guess on these, setting all the tick
params to true allocated extra margin space?), so that's why
the small example dotplot PNG changed apparently. (the e. coli
example seemed to stay the same.)

also, i spent some time investigating adding ticks / tick labels
to the e. coli example, but the code to do that was pretty sloppy
and involved -- out of scope of the tutorial. in the future,
it'd be good to add an option to the viz() functions that adds
ticks and auto-detects order of magnitude, etc.