TST/DOC/DEP: Move _fill_match_cells() doctest
See the comments. Basically, this was the least bad solution I could
think of to handle numpy 2's new string representations of
np.int32/etc types.

We COULD cast the entries in the suffix arrays from np.int32 to int,
but that doesn't actually make a difference (np.int32(5) == 5; note
that all of the other tests were still passing in the Python 3.11
numpy 2 environment) and will slow things down. I think the better
long-term solution (rather than trying to move heaven and earth
to preserve a cross-platform doctest here) is just moving the
doctest to a unit test, as discussed in the context of python 2->3
migration here: https://stackoverflow.com/q/13473971