PERF: add very basic memory profiling to b ntbk
for some reason the 100M x 100M test stil fails - not sure why.

i tried out running the notebook in both firefox and chrome, so that
confirms that this is not a browser issue. next up, try running old
code and see if the memory numbers look different and/or if the
100M x 100M test works